app.js and ChoreChart.js hooked up. First step
